web-wallet: Update duskToLux and luxToDusk
Resolves #2736, Resolves #2737
