+++ web-wallet/src/lib/contracts/deductLuxFeeFrom.js
@@ -11,6 +11,6 @@ import { duskToLux, luxToDusk } from "$lib/dusk/currency";
  * @returns {number}
  */
 const deductLuxFeeFrom = (duskAmount, luxFee) =>
-  +luxToDusk(duskToLux(duskAmount) - luxFee).toFixed(9);
+  +luxToDusk(duskToLux(duskAmount) - BigInt(luxFee)).toFixed(9);
 
 export default deductLuxFeeFrom;
