+++ web-wallet/src/lib/dusk/currency/__tests__/duskToLux.spec.js
@@ -4,7 +4,7 @@ import { duskToLux } from "..";
 
 describe("duskToLux", () => {
   it("should convert an amount in Dusk to Lux", () => {
-    expect(duskToLux(1)).toBe(1e9);
-    expect(duskToLux(3_456_789.012)).toBe(3_456_789_012_000_000);
+    expect(duskToLux(1)).toBe(BigInt(1e9));
+    expect(duskToLux(3_456_789.012)).toBe(BigInt(3_456_789_012_000_000));
   });
 });
